requirements.txt
# python 3.11

# pipdeptree==2.26.1                      # check dependency between packages
python-dotenv==1.1.0                    # export variable from .env
# ipykernel==6.29.5                       # jupyter notebook
gradio==5.33.2                          # for a quick web ui

# lang stuff
langchain==0.3.25
langgraph==0.4.8
langchain-google-genai==2.1.5           # call google-genai via builtin API
langchain-community==0.3.25             # load data, call pypdf via builtin API
langchain-chroma==0.2.4                 # vector db

# tool
pypdf==5.6.0                            # read pdf
protobuf==3.20.2                        # avoid conflict with chroma db
langchain-tavily==0.2.3                 # search tool (on the internet)

pytest==8.4.1                           # pytest -v
